check dir exist

+++ sync.sh
@@ -6,18 +6,36 @@ do
 		o)  #orientation
                         if [ $OPTARG = "backup" ]
                         then
+                                #vim
                                 cp ~/.vimrc                       ./
                                 cp -r ~/.vim/                     ./
+                                #shadowsocks
+                                if [ ! -e ./.ss.d/ ]
+                                then 
+                                        mkdir ./.ss.d
+                                if
                                 cp ~/.ss.pac                      ./.ss.d/
                                 cp /etc/shadowsocks/.ss.conf.json ./.ss.d/
+                                #conda
+                                if [ ! -e .conda ]
+                                then
+                                        mkdir .conda
+                                if
                                 source activate mltoolchain
                                 conda env export -f               ./.conda/mltoolchain.yml
                         elif [ $OPTARG = "reduct" ]
                         then
+                                #vim
                                 cp .vimrc              ~/
                                 cp -r .vim/            ~/
+                                #shadowsocks
                                 cp .ss.d/.ss.pac       ~/
-                                cp .ss.d/.ss.conf.json /etc/shadowsocks/
+                                if [ -e /etc/shadowsocks/ ] 
+                                then
+                                        cp .ss.d/.ss.conf.json /etc/shadowsocks/
+                                else
+                                        echo "null /ect/shadowsocks!"
+                                if
                         else
                                 echo "Unknown argument $OPTARG"
                         fi
